+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix wrong error occurrence when there is undefined participant in component.
 - Allow to generate a dir with all sub tucs of a multi-tuc file.
 - Enhance an output for generating tuc
+- Generate method arguments of service-method-call as multiline, when there are more than one
 
 ## 1.1.0 - 2020-09-21
 - Allow Modules, Functions, etc in Domain files.